Add Meziantou.Analyzer to test projects
Catches real issues at build time: MA0048 (file name must match type name)
caught ThrowingAppFactory defined in the wrong file, which is now moved to
ThrowingAppFactory.cs. MA0074 suppressed — xunit's Assert.Contains(string, string)
has no StringComparison overload so the rule is a false positive for tests.
