+++ dotnet-minimal-api/template/tests/MyMinimalWebApp.Api.IntegrationTests/Middleware/ExceptionMiddlewareTests.cs
@@ -26,16 +26,3 @@ public class ExceptionMiddlewareTests : IClassFixture<ThrowingAppFactory>
         Assert.Contains("500", body);
     }
 }
-
-public class ThrowingAppFactory : WebApplicationFactory<Program>
-{
-    protected override void ConfigureWebHost(IWebHostBuilder builder)
-    {
-        builder.Configure(app =>
-        {
-            app.UseMiddleware<MyMinimalWebApp.Api.Middleware.ExceptionMiddleware>();
-            app.Map("/throw", throwApp =>
-                throwApp.Run(_ => throw new InvalidOperationException("Test unhandled exception")));
-        });
-    }
-}
