+++ dotnet-minimal-api/template/tests/MyMinimalWebApp.Api.UnitTests/MyMinimalWebApp.Api.UnitTests.csproj
@@ -7,6 +7,7 @@
   <ItemGroup>
     <PackageReference Include="Bogus" />
     <PackageReference Include="coverlet.collector" />
+    <PackageReference Include="Meziantou.Analyzer" PrivateAssets="all" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="NSubstitute" />
     <PackageReference Include="xunit" />
