+++ dotnet-minimal-api/template/tests/MyMinimalWebApp.Api.UnitTests/Services/ItemServiceTests.cs
@@ -103,3 +103,4 @@ public class ItemServiceTests
         Assert.False(deleted);
     }
 }
+
